added listing page, fixed a few bugs

+++ src/services/SpendingServices.ts
@@ -4,13 +4,14 @@ import { spendingRef } from "../functions/Firebase"
 import { numToDay, numToMonth } from "../helpers/Generichelper"
 
 
-export const createSpendingMonth = (uid: string, onSucess: (docID: string) => void, onError: (error: string) => void) => {
+export const createSpendingMonth = (uid: string, initialAmount: number, onSuccess: (docID: string) => void, onError: (error: string) => void) => {
   spendingRef.add({
     user_id: uid,
+    amount: initialAmount,
     month: numToMonth(moment().toDate().getMonth()),
     year: moment().toDate().getFullYear(),
   }).then(docRef => {
-    onSucess(docRef.id);
+    onSuccess(docRef.id);
   }).catch(err => {
     onError(err);
   })
@@ -31,7 +32,7 @@ export const checkMonthExist = (uid: string, month: string, year: number, ifExis
     })
 }
 
-export const createSpendingTab = (uid: string, data: { amount: string, description: string }, onSucess: (docID: string) => void, onError: (error: string) => void) => {
+export const createSpendingTab = (uid: string, data: { amount: string, description: string }, onSuccess: (docID: string) => void, onError: (error: string) => void) => {
   const { amount, description } = data;
 
   checkMonthExist(
@@ -41,20 +42,33 @@ export const createSpendingTab = (uid: string, data: { amount: string, descripti
       //if exist
       spendingRef
         .doc(docID)
-        .collection(TABS)
-        .add({
-          amount: Number(amount),
-          description: description,
-          created_date: `${moment().toDate().getDate()}/${moment().toDate().getMonth() + 1}/${moment().toDate().getFullYear()}`,
-          created_day: numToDay(moment().toDate().getDay()),
-          created_at: moment().toDate()
+        .get()
+        .then(snapshot => {
+          if (snapshot.data()) {
+            spendingRef.doc(docID).set(
+              { amount: Number(amount) + snapshot.data()!.amount },
+              { merge: true }
+            ).then(() => {
+              spendingRef
+                .doc(docID)
+                .collection(TABS)
+                .add({
+                  amount: Number(amount),
+                  description: description,
+                  created_date: `${moment().toDate().getDate()}/${moment().toDate().getMonth() + 1}/${moment().toDate().getFullYear()}`,
+                  created_day: numToDay(moment().toDate().getDay()),
+                  created_at: moment().toDate()
+                })
+                .then(() => {
+                  onSuccess(docID);
+                })
+            })
+          }
         })
-        .then(() => {
-          onSucess(docID);
-        })
+
     }, () => {
       //if not exist
-      createSpendingMonth(uid, (docID) => {
+      createSpendingMonth(uid, Number(amount), (docID) => {
         spendingRef
           .doc(docID)
           .collection(TABS)
@@ -66,7 +80,7 @@ export const createSpendingTab = (uid: string, data: { amount: string, descripti
             created_at: moment().toDate()
           })
           .then(() => {
-            onSucess(docID);
+            onSuccess(docID);
           })
       }, (err) => {
         onError(err);
@@ -75,12 +89,26 @@ export const createSpendingTab = (uid: string, data: { amount: string, descripti
     })
 }
 
-export const deleteSpendingCard = (collectionID: string, tabID: string, onSuccess: () => void, onError: (err: string) => void) => {
-  spendingRef.doc(collectionID).collection(TABS).doc(tabID).delete().then(() => {
-    onSuccess();
-  }).catch((error) => {
-    onError(error);
-    console.error("Error removing document: ", error);
-  });
+export const deleteSpendingCard = (collectionID: string, tabID: string, amount: number, onSuccess: () => void, onError: (err: string) => void) => {
+  spendingRef
+    .doc(collectionID)
+    .get()
+    .then(snapshot => {
+      if (snapshot.data()) {
+        spendingRef
+          .doc(collectionID)
+          .set({
+            amount: snapshot.data()!.amount - amount
+          }, { merge: true })
+          .then(() => {
+            spendingRef.doc(collectionID).collection(TABS).doc(tabID).delete().then(() => {
+              onSuccess();
+            }).catch((error) => {
+              onError(error);
+              console.error("Error removing document: ", error);
+            });
+          })
+      }
+    })
 }
